refactor(columnar): extract relation-plan dispatch
